+++ Examples/Run/Geant4/HepMC/EventRecordingExample.cpp
@@ -78,5 +78,5 @@ int main(int argc, char* argv[]) {
       std::move(hepMC3WriterConfig), logLevel));
 
   // Run
-  return sequencer.run();
+  sequencer.run();
 }
